+++ command/ca/certificate.go
@@ -4,7 +4,6 @@ import (
 	"strings"
 
 	"github.com/pkg/errors"
-	"github.com/smallstep/cli/crypto/pemutil"
 	"github.com/smallstep/cli/flags"
 	"github.com/smallstep/cli/token"
 	"github.com/smallstep/cli/utils/cautils"
@@ -12,6 +11,7 @@ import (
 	"go.step.sm/cli-utils/command"
 	"go.step.sm/cli-utils/errs"
 	"go.step.sm/cli-utils/ui"
+	"go.step.sm/crypto/pemutil"
 )
 
 func certificateCommand() cli.Command {
@@ -186,8 +186,10 @@ Use the '--permanent-identifier' flag multiple times to set more than one.`,
 			flags.Offline,
 			flags.PasswordFile,
 			consoleFlag,
+			flags.KMSUri,
 			flags.X5cCert,
 			flags.X5cKey,
+			flags.X5cChain,
 			flags.NebulaCert,
 			flags.NebulaKey,
 			acmeFlag,