+++ command/ca/sign.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/smallstep/certificates/api"
-	"github.com/smallstep/cli/crypto/pemutil"
 	"github.com/smallstep/cli/flags"
 	"github.com/smallstep/cli/token"
 	"github.com/smallstep/cli/utils/cautils"
@@ -14,6 +13,7 @@ import (
 	"go.step.sm/cli-utils/command"
 	"go.step.sm/cli-utils/errs"
 	"go.step.sm/cli-utils/ui"
+	"go.step.sm/crypto/pemutil"
 )
 
 func signCertificateCommand() cli.Command {
@@ -125,8 +125,10 @@ $ step ca sign foo.csr foo.crt \
 			flags.Offline,
 			flags.PasswordFile,
 			consoleFlag,
+			flags.KMSUri,
 			flags.X5cCert,
 			flags.X5cKey,
+			flags.X5cChain,
 			flags.NebulaCert,
 			flags.NebulaKey,
 			acmeFlag,