+++ CHANGELOG.md
@@ -22,8 +22,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added support for configuring ACME device-attest-01 challenges.
 - Added support for ACME device-attest-01 challenges with YubiKeys.
 - Added support for SUSE13 and upwards for `step certificate install`.
+- Added support for printing [Sigstore](https://www.sigstore.dev/) certificate details to `step certificate inspect`
+- Added `--acme` flag to the `step ca init` command, to create a default ACME provisioner when initializing a CA.
+- Added `--remote-management` flag to the `step ca init` command, which enables Remote Management of the CA using the Admin API.
 ### Changed
 - New OAuth success page with color.
+- Added `x5c-roots` as alias for `x5c-root` flag.
 
 ## [0.22.0] - 2022-08-25
 ### Added
@@ -161,11 +165,11 @@ to the value of provisioner-password-file flag.
 
 ## [0.0.2]
 ### Added
-- `--bundle` flag to cert/inspect for inpecting all the full chain or bundle
+- `--bundle` flag to cert/inspect for inspecting all the full chain or bundle
 given a path. Default behavior is unchanged; only inspect the first (leaf)
 certificate.
 - distribution.md with documentation on how to create releases.
-- travis build and upload artifacts to Github Releases on tagged pushes.
+- travis build and upload artifacts to GitHub Releases on tagged pushes.
 - logging of invalid http requests to the oauth server
 ### Changed
 - default PEM format encryption alg AES128 -> AES256